Creando todos os endpoints de categorias

+++ src/ControleFinanceiro.Repository/Interfaces/IRepository.cs
@@ -11,7 +11,7 @@ namespace ControleFinanceiro.Repository.Interfaces
     {
             void Create(T entity);
             T Find(params object[] keyValues);
-            void Edit(T entity, int id);
+            void Edit(T entity, long id);
             void Remove(int[] keyValues);
             void Remove(T entity);
             List<T> List(string foreignKey = null);
